helm app overview refactoring
