+++ src/components/app/Overview/Overview.tsx
@@ -30,6 +30,7 @@ import {
     noop,
     StyledRadioGroup as RadioGroup,
     EditableTextArea,
+    APP_TYPE,
 } from '@devtron-labs/devtron-fe-common-lib'
 import ReactGA from 'react-ga4'
 import { handleUTCTime, importComponentFromFELibrary } from '../../common'
@@ -72,8 +73,8 @@ export default function AppOverview({ appMetaInfo, getAppMetaInfoRes, filteredEn
     const isUpdateDependencyModalOpen =
         activeTab === OVERVIEW_TABS.DEPENDENCIES && searchParams.get(MODAL_STATE.key) === MODAL_STATE.value
     const config = getAppConfig(appType)
-    const isJobOverview = appType === 'job'
-    const isHelmChart = appType === 'helm-chart'
+    const isJobOverview = appType === APP_TYPE.JOB
+    const isHelmChart = appType === APP_TYPE.HELM_CHART
     // For helm the appId from the params is the installed appId and not the actual id of the app
     const appId = isHelmChart ? `${appMetaInfo.appId}` : appIdFromParams
     const [isLoading, setIsLoading] = useState(true)
@@ -307,7 +308,7 @@ export default function AppOverview({ appMetaInfo, getAppMetaInfoRes, filteredEn
                             {createdBy}
                         </div>
                     </div>
-                    {appType === 'app' && gitMaterials.length > 0 && (
+                    {appType === APP_TYPE.DEVTRON_APPS && gitMaterials.length > 0 && (
                         <div>
                             <div className="fs-13 fw-4 lh-20 cn-7 mb-4">Code source</div>
                             <div className="flexbox-col dc__gap-4">
