+++ src/components/app/types.ts
@@ -15,7 +15,7 @@
  */
 
 import React, { ReactNode } from 'react'
-import { ACTION_STATE, DeploymentAppTypes, TagType, Teams, PodMetadatum } from '@devtron-labs/devtron-fe-common-lib'
+import { ACTION_STATE, DeploymentAppTypes, TagType, Teams, PodMetadatum, APP_TYPE } from '@devtron-labs/devtron-fe-common-lib'
 import { RouteComponentProps } from 'react-router'
 import { AppEnvironment } from '../../services/service.types'
 import { DeploymentStatusDetailsBreakdownDataType, ErrorItem } from './details/appDetails/appDetails.type'
@@ -499,7 +499,7 @@ export interface AppOverviewProps {
      *
      * @default 'app'
      */
-    appType: 'job' | 'app' | 'helm-chart'
+    appType: APP_TYPE.JOB | APP_TYPE.DEVTRON_APPS | APP_TYPE.HELM_CHART
 }
 
 export interface OverviewConfig {
